chore(all): rename embarrasing typo.
deary me!
